django-react/requirements.txt
asgiref
certifi
charset-normalizer
click
colorama
Django
django-cors-headers
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt
idna
joblib
nltk
numpy
pandas
PyJWT
python-dateutil
pytz
regex
requests
six
sqlparse
tqdm
tzdata
urllib3
vaderSentiment
